backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
farm-haystack==1.20.0
haystack-integrations==0.1.0
pgvector==0.2.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pdfkit==1.0.0
requests==2.31.0
bcrypt==4.0.1
